reduce differences to upstream version

+++ teensy3/pins_teensy.c
@@ -762,7 +762,7 @@ void analogWrite(uint8_t pin, int val)
 		digitalWrite(pin, LOW);
 		pinMode(pin, OUTPUT);	// TODO: implement OUTPUT_LOW
 		return;
-	} else if ((uint32_t) val >= max) {
+	} else if (val >= max) {
 		digitalWrite(pin, HIGH);
 		pinMode(pin, OUTPUT);	// TODO: implement OUTPUT_HIGH
 		return;
